Added authorization, fixed the models

+++ src/main/java/ea/sof/ms_answers/controller/AnswerController.java
@@ -3,9 +3,12 @@ package ea.sof.ms_answers.controller;
 import ea.sof.ms_answers.entity.AnswerEntity;
 import ea.sof.ms_answers.model.AnswerReqModel;
 import ea.sof.ms_answers.repository.AnswerRepository;
+import ea.sof.ms_answers.service.AuthService;
 import ea.sof.shared.models.Answer;
 import ea.sof.shared.models.Response;
+import ea.sof.shared.models.TokenUser;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,12 +18,14 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @RestController
-@RequestMapping("/api/answers")
+@RequestMapping("/answers")
 public class AnswerController {
     @Autowired
     AnswerRepository answerRepository;
+    @Autowired
+    AuthService authService;
 
-    @GetMapping
+    /*@GetMapping
     public ResponseEntity<?> getAllAnswers() {
         List<AnswerEntity> answerEntities = answerRepository.findAll();
         List<Answer> answers = answerEntities.stream().map(ans -> ans.toAnswerModel()).collect(Collectors.toList());
@@ -29,7 +34,7 @@ public class AnswerController {
         response.getData().put("answers", answers);
 
         return ResponseEntity.ok(response);
-    }
+    }*/
 
     @GetMapping("/question/{questionId}")
     public ResponseEntity<?> getAllAnswersByQuestionId(@PathVariable("questionId") String questionId) {
@@ -46,76 +51,129 @@ public class AnswerController {
     public ResponseEntity<?> getAnswerById(@PathVariable("answerId") String answerId) {
         AnswerEntity answerEntity = answerRepository.findById(answerId).orElse(null);
         if(answerEntity == null) {
-            return ResponseEntity.status(404).body(new Response(false, "No match found"));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Response(false, "No match found"));
         }
         Response response = new Response(true, "");
-        response.getData().put("answer", answerEntity);
+        response.getData().put("answer", answerEntity.toAnswerModel());
         return ResponseEntity.ok(response);
     }
 
     @PostMapping("/{questionId}")
-    public ResponseEntity<?> createAnswer(@RequestBody @Valid AnswerReqModel answerReqModel, @PathVariable("questionId") String questionId) {
+    public ResponseEntity<?> createAnswer(@RequestBody @Valid AnswerReqModel answerReqModel, @PathVariable("questionId") String questionId, @RequestHeader("Authorization") String token) {
+
+        //Check if request is authorized
+        Response authCheckResp = isAuthorized(token);
+        if (!authCheckResp.getSuccess()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new Response(false, "Invalid Token"));
+        }
+        TokenUser decodedToken = (TokenUser) authCheckResp.getData().get("decoded_token");
+
         AnswerEntity answerEntity = new AnswerEntity(answerReqModel);
         answerEntity.setQuestionId(questionId);
-        //todo: answerEntity.setUserId();
+        answerEntity.setUserId(decodedToken.getUserId().toString());
+
         Response response = new Response(true, "Answer has been created");
         answerEntity = answerRepository.save(answerEntity);
-        response.getData().put("answer", answerEntity);
-        return ResponseEntity.status(201).body(response);
+        response.getData().put("answer", answerEntity.toAnswerModel());
+        return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
     @PatchMapping("/{answerId}/upvote")
-    public ResponseEntity<?> upVote(@PathVariable("answerId") String answerId) {
+    public ResponseEntity<?> upVote(@PathVariable("answerId") String answerId, @RequestHeader("Authorization") String token) {
+
+        //Check if request is authorized
+        Response authCheckResp = isAuthorized(token);
+        if (!authCheckResp.getSuccess()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new Response(false, "Invalid Token"));
+        }
+
         AnswerEntity answerEntity = answerRepository.findById(answerId).orElse(null);
         if(answerEntity == null) {
-            return ResponseEntity.status(404).body(new Response(false, "No match found"));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Response(false, "No match found"));
         }
         answerEntity.upvote();
         answerEntity = answerRepository.save(answerEntity);
         Response response = new Response(true, "Answer upVoted");
-        response.getData().put("answer", answerEntity);
+        response.getData().put("answer", answerEntity.toAnswerModel());
         return ResponseEntity.ok(response);
     }
 
     @PatchMapping("/{answerId}/downvote")
-    public ResponseEntity<?> downVote(@PathVariable("answerId") String answerId) {
-        //
+    public ResponseEntity<?> downVote(@PathVariable("answerId") String answerId, @RequestHeader("Authorization") String token) {
+
+        //Check if request is authorized
+        Response authCheckResp = isAuthorized(token);
+        if (!authCheckResp.getSuccess()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new Response(false, "Invalid Token"));
+        }
+
         AnswerEntity answerEntity = answerRepository.findById(answerId).orElse(null);
         if(answerEntity == null) {
-            return ResponseEntity.status(404).body(new Response(false, "No match found"));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Response(false, "No match found"));
         }
         answerEntity.downvote();
         answerEntity = answerRepository.save(answerEntity);
         Response response = new Response(true, "Answer downVoted");
-        response.getData().put("answer", answerEntity);
+        response.getData().put("answer", answerEntity.toAnswerModel());
         return ResponseEntity.ok(response);
     }
 
     @PatchMapping("/{answerId}")
-    public ResponseEntity<?> updateAnswer(@PathVariable("answerId") String answerId, @RequestBody @Valid AnswerReqModel answerReqModel) {
+    public ResponseEntity<?> updateAnswer(@PathVariable("answerId") String answerId, @RequestBody @Valid AnswerReqModel answerReqModel, @RequestHeader("Authorization") String token) {
+
+        //Check if request is authorized
+        Response authCheckResp = isAuthorized(token);
+        if (!authCheckResp.getSuccess()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new Response(false, "Invalid Token"));
+        }
+
         AnswerEntity answerEntity = answerRepository.findById(answerId).orElse(null);
         if(answerEntity == null) {
-            return ResponseEntity.status(404).body(new Response(false, "No match found"));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Response(false, "No match found"));
         }
         AnswerEntity newAnswerEntity = new AnswerEntity(answerReqModel);
         answerEntity.setBody(newAnswerEntity.getBody());
         answerEntity.setLastEdited(newAnswerEntity.getCreated());
         answerEntity = answerRepository.save(answerEntity);
         Response response = new Response(true, "Answer updated");
-        response.getData().put("answer", answerEntity);
+        response.getData().put("answer", answerEntity.toAnswerModel());
         return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("/{answerId}")
-    public ResponseEntity<?> deleteAnswer(@PathVariable("answerId") String answerId) {
+    public ResponseEntity<?> deleteAnswer(@PathVariable("answerId") String answerId, @RequestHeader("Authorization") String token) {
+
+        //Check if request is authorized
+        Response authCheckResp = isAuthorized(token);
+        if (!authCheckResp.getSuccess()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new Response(false, "Invalid Token"));
+        }
+
         AnswerEntity answerEntity = answerRepository.findById(answerId).orElse(null);
         if(answerEntity == null) {
-            return ResponseEntity.status(404).body(new Response(false, "No match found"));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new Response(false, "No match found"));
         }
         answerRepository.delete(answerEntity);
 
         Response response = new Response(true, "Answer deleted");
-        response.getData().put("answer", answerEntity);
+        response.getData().put("answer", answerEntity.toAnswerModel());
         return ResponseEntity.ok(response);
+    }
+
+    private Response isAuthorized(String authHeader) {
+        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
+            return new Response(false, "Invalid token");
+        }
+        try {
+            ResponseEntity<Response> result = authService.validateToken(authHeader);
+
+            if (!result.getBody().getSuccess()) {
+                return new Response(false, "Invalid token");
+            }
+            return result.getBody();
+
+        }catch (Exception e){
+            return new Response(false, "exception", e);
+        }
     }
 }
